bot_telegram/requirements.txt
python-telegram-bot
supabase>=2.10.0
openai
python-dotenv
pydub
requests
websockets>=13.0
httpx>=0.28.0
